refactor: single VBAHostFile base for the three host facades
ExcelFile, WordFile, and PowerPointFile were ~460-line near-copies;
the safety-critical save() pipeline and the issue #1 class
normalization had to be hand-synchronized across three files (drift
had already crept in: ExcelFile carried a dead _zip_bytes attribute
the others lacked). All shared behavior now lives once in
pyopenvba._host.VBAHostFile, parameterized by container extensions,
the vbaProject.bin entry path, and message wording; the subclasses
keep their create_new(), constants, and docstrings. Public API and
error-message text are unchanged.

Two behavior changes ride along, both covered by new tests:

- save() iterates pending adds and deletes in sorted order, making
  multi-add output byte-deterministic across processes (Python
  randomizes set order via string hashing; tests/test_determinism.py
  proves stability across PYTHONHASHSEED values).
- PROJECTwm is addressed at the project root, where [MS-OVBA] 2.2.1
  places it (verified across all five live container layouts). The
  old storage-qualified lookup only worked because CFB storage
  lookups ignored storage boundaries.

Verified byte-identical against the previous implementation: a
25-case save matrix (noop/edit/add/rename/delete over the xlsm, xlsb,
xls, docm, and pptm live fixtures) hashes identically before and
after.
